Basic repo that contains a TS compiler error and a custom watch task with a problemMatcher that doesn't match expectations with respect to Problems view.
If owner is set to typescript, after opening the file by clicking on the error in Problems view there is still only one problem in the list (but it replaces the contents of the list with the $tsc-watch problemMatcher). Closing the file clears the error from the Problems view, even though the error is still present. If I re-run the running build task, the error returns to the Problems view (with the format of the custom problemMatcher).

If the owner is set to a unique string, i.e. "erin", after opening the file by clicking the link in the Problems view, there are two problems in the list (1 that matches the $tsc-watch problemMatcher and 1 that matches the problemMatcher as defined in the Custom watch task).

After closing the file without fixing the error, the $tsc-watch problem disappears from the Problems view.

After fixing the error, the $tsc-watch error disappears from the Problems view but the one from the Custom watch problemMatcher does not disappear from the Problems view. Closing the file doesn't affect it. It only goes away when I restart the Custom watch task.